requirements.txt
# Core Dependencies
anthropic>=0.39.0
pyboy>=2.2.1
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0

# Image Processing
opencv-python>=4.8.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
colorlog>=6.8.0

# Data Handling
pandas>=2.0.0
jsonschema>=4.20.0

# Optional: Web Interface (for future dashboard)
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
